% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseInfo.R
\name{collapseInfo}
\alias{collapseInfo}
\title{Collapse list of enviornment information into a single data frme}
\usage{
collapseInfo(info_list = list())
}
\arguments{
\item{info_list}{List with environment information}
}
\description{
Collapse list of enviornment information into a single data frme
}
