% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envDocument.R
\docType{package}
\name{envDocument-package}
\alias{envDocument}
\alias{envDocument-package}
\title{Document the environment under which an analysis was performed}
\description{
Provides a function \code{\link{env_doc}} that collects information on the
script, environment and system that performed an analysis.
}
\details{
Returns a data frame with information on:  

 - R version  

 - Attacheded packages and version   
 
 - OS, version and user
 
 - Path and modification time of the calling script   
 
  - Git commit and tag info (optional, requires 
      \code{git2r})

Results are collected in a standardized data frame that can be queried or
pretty-printed using \code{\link[knitr]{kable}} or similar methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dgJacks0n/envDocument}
}

}
\author{
\strong{Maintainer}: Donald Jackson \email{14615921+dgJacks0n@users.noreply.github.com}

Authors:
\itemize{
  \item Donald Jackson \email{14615921+dgJacks0n@users.noreply.github.com}
}

}
