% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_environment.R
\name{get_gitInfo}
\alias{get_gitInfo}
\title{get_gitInfo}
\usage{
get_gitInfo(scriptPath = "")
}
\arguments{
\item{script}{(optional, defaults to get_scriptPath())}
}
\description{
Get git information from repository
}

