% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/document_environment.R
\name{get_packageinfo}
\alias{get_packageinfo}
\title{get_packageinfo}
\usage{
get_packageinfo()
}
\arguments{
\item{none}{}
}
\description{
Get information on attached packages, return as a data frame
}
\examples{
packages <- get_packageinfo
}

