% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/document_environment.R
\name{get_rversion}
\alias{get_rversion}
\title{get_rversion}
\usage{
get_rversion()
}
\arguments{
\item{none}{}
}
\description{
Get selected info on R version
}
\examples{
Rversion <- get_rversion()
}

