% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scriptinfo.R
\name{get_scriptinfo}
\alias{get_scriptinfo}
\title{Get information about the  calling script}
\usage{
get_scriptinfo()
}
\description{
\code{get_scriptinfo} retrieves the  file path and  modification time
for the calling script.
}
\details{
Note:
Currently this works if the script was called via source(),
knitr::spin() or Rstudio's _compile notebook_ (ctrl-shift-K).
It does not work if the script was called via 'R -f' or 'Rscript'.
In those cases NAs are returned for the path and mod. time.
A later version should capture additional cases.
}
\examples{
scriptinfo <- get_scriptinfo()
}

