% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/document_environment.R
=======
% Please edit documentation in R/get_scriptpath.R
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae
\name{get_scriptpath}
\alias{get_scriptpath}
\title{Get the path of the calling script}
\usage{
get_scriptpath(absolute = TRUE)
}
<<<<<<< HEAD
=======
\arguments{
\item{absolute}{Return absolute path (absolute = TRUE, default) or path relative
to working directory}
}
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae
\description{
\code{get_scriptpath} returns the full path of the script that called this function (if any)
or NULL if path is not available
}
\examples{
mypath <- get_scriptpath()

}

