% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/document_environment.R
\name{get_sysinfo}
\alias{get_sysinfo}
\title{Get system information
  
Get system information: OS, hostname, userid, working directory as a data frame}
=======
% Please edit documentation in R/get_sysinfo.R
\name{get_sysinfo}
\alias{get_sysinfo}
\title{Get system information.}
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae
\usage{
get_sysinfo()
}
\description{
<<<<<<< HEAD
Get system information
  
Get system information: OS, hostname, userid, working directory as a data frame
=======
\code{get_sysinfo} returns the OS, hostname, userid and working directory
 for the current analysis as a data frame
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae
}
\examples{
get_sysinfo()

}

