% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/document_environment.R
\name{get_thispackageinfo}
\alias{get_thispackageinfo}
\title{get_thispackageinfo}
\usage{
get_thispackageinfo(thisPackage)
}
\description{
extract selected info on individual packages.  Called by get_packageinfo
}

