% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominoInfo.R
\name{getDominoInfo}
\alias{getDominoInfo}
\title{Get information about a Domino run from environment variables}
\usage{
getDominoInfo()
}
\description{
Get information about a Domino run from environment variables
}
