% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGitinfo.R
\name{getGitinfo}
\alias{getGitinfo}
\title{Get git repository information for a script}
\usage{
getGitinfo(scriptpath = "")
}
\arguments{
\item{scriptpath}{Path to script (optional, defaults to get_scriptPath())}
}
\description{
\code{getGitinfo} locates the git repository for the calling script
(if any) and retrieves relevant informaiton such as last commit, status
(e.g. modified since last commit) and tag (if any).
}
\details{
Requires that \code{git2r} be installed.
}
\examples{
git_info <- getGitinfo()

}

