% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageInfo.R
\name{getPackageInfo}
\alias{getPackageInfo}
\title{Get information on attached packages}
\usage{
getPackageInfo()
}
\description{
\code{getPackageInfo} returns information on attached packages
including name, version and source.
}
\examples{
packages <- getPackageInfo()
}
