% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRenvStatus.R
\name{getRenvStatus}
\alias{getRenvStatus}
\title{Check whether \code{renv} is in use for this project.  
  If it is, get project status.
  Requires that \code{renv} be installed and active.}
\usage{
getRenvStatus(project_root = NA)
}
\arguments{
\item{project_root}{Project root directory (optional); defaults to value
from \code{here::here()}}
}
\description{
Check whether \code{renv} is in use for this project.  
  If it is, get project status.
  Requires that \code{renv} be installed and active.
}
