% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepo.R
\name{getRepo}
\alias{getRepo}
\title{Get git repository for a file}
\usage{
getRepo(testPath)
}
\arguments{
\item{testPath}{Path to a file to find/check repo}
}
\description{
\code{getRepo} looks the up the repository for a file and 
makes sure path is tracked in repo.  Repositories are searched
from the file directory up using 
\code{\link{git2r::discover_repository}}
}

