% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptPath.R
\name{getScriptPath}
\alias{getScriptPath}
\alias{get_scriptpath}
\title{Get the path of the calling script}
\usage{
getScriptPath(absolute = TRUE)
}
\arguments{
\item{absolute}{Return absolute path (absolute = TRUE, default) or path relative
to working directory (FALSE)}
}
\description{
\code{getScriptPath} returns the full path of the script that called this function (if any)
or NULL if path is not available
}
\examples{
\dontrun{
mypath <- getScriptPath()
}

}
