% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSysInfo.R
\name{getSysInfo}
\alias{getSysInfo}
\title{Get system information.}
\usage{
getSysInfo()
}
\description{
\code{getSysInfo} returns the OS, hostname, userid and working directory
 for the current analysis as a data frame
}
\examples{
getSysInfo()

}
