% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_environment.R
\name{getTag}
\alias{getTag}
\title{getTag}
\usage{
getTag(repo)
}
\arguments{
\item{repo}{}
}
\description{
Return git tag information if relavant
Requires that repostiory have tags and that sha for last tag
and last commit match.
}

