% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoNotFound.R
\name{infoNotFound}
\alias{infoNotFound}
\title{Return a consistent failure result when something is not available}
\usage{
infoNotFound()
}
\description{
Return a consistent failure result when something is not available
}
