\name{print_packageinfo}
\alias{print_packageinfo}
\title{
print_packageinfo: function to list included packages with version information
}
\description{
Prints a list of all currently included packages with version information.
Designed to facilitate code reproducibility.  Called by env_doc if
the 'packages' argument is TRUE (default).

Note: any packages included AFTER this function is called won't be listed.
}
\usage{
print_packageinfo()
}
\details{
}
\value{
}
\references{
}
\author{
Donald Jackson
Bristol-Myers Squibb
donald.jackson@bms.com
}
\note{
}


\seealso{
}
\examples{
print_packageinfo()
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
