% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/document_environment.R
=======
% Please edit documentation in R/env_doc.R
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae
\name{env_doc}
\alias{env_doc}
\title{Report the working environment in which an analysis was performed.}
\usage{
env_doc(output = c("return", "print", "table"), system = TRUE,
  version = TRUE, packages = TRUE, script = TRUE, git = TRUE)
}
\arguments{
<<<<<<< HEAD
\item{output}{How should output be handled? return: return as a data frame (default);
print: print to stdout
@param system Include OS info from get_sysinfo()? Default TRUE
@param version Include R version?  Default TRUE
@param packages Include packages with repository and version from get_packageinfo()? Default TRUE
@param script Include script path and modification time from get_scriptinfo()? Default TRUE

@examples
env_doc("print") # print information to stdout
info <- env_doc() # return information as a consolidated data frame}
=======
\item{output}{How should output be handled? return: return as a 
data frame (default); print: print to stdout; table: pretty-print
using knitr::kable (requires package knitr)}

\item{system}{Include OS info from \code{\link{get_sysinfo}}? Boolean, 
default TRUE}

\item{version}{Include R version from \code{\link{get_rversion}}?
Boolean, default TRUE}

\item{packages}{Include attached packages with repository and version from 
\code{\link{get_packageinfo}}? Boolean, default TRUE}

\item{script}{Include script path and modification time from 
\code{\link{get_scriptinfo}}? Boolean, default TRUE}

\item{git}{Include git repository information from \code{\link{get_gitInfo}}
(note: requires git2r)?  Boolean, default TRUE}
}
\value{
If outpt = return (default) :A data frame with columns for information 
  type, variable name and value.  NULL for output = print or table
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae
}
\description{
\code{env_doc} prints or returns a data frame with the following information:  
- System version (OS, version, user, working directory)
}
\details{
- R version

- Package names and versions

- Top-level script name and modification time

- Git hash, status and tag (if any; requires package git2r)
}
<<<<<<< HEAD
=======
\examples{
 env_doc("print") # print information to stdout
 info <- env_doc() # return information as a consolidated data frame
}
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae

