% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_doc.R
\name{env_doc}
\alias{env_doc}
\title{env_doc
\code{env_doc} reports the working environment}
\usage{
env_doc(output = c("return", "print"), system = TRUE, version = TRUE,
  packages = TRUE, script = TRUE, git = TRUE)
}
\arguments{
\item{output}{How should output be handled? return: return as a 
data frame (default); print: print to stdout}

\item{system}{Include OS info from get_sysinfo()? Default TRUE}

\item{version}{Include R version?  Default TRUE}

\item{packages}{Include packages with repository and version from get_packageinfo()? Default TRUE}

\item{script}{Include script path and modification time from get_scriptinfo()? Default TRUE}

\item{git}{Include git repository information? from get_gitInfo (note: requires git2r)?  Default TRUE}
}
\description{
Returns a data frame with information about the script, system
and environment
}
\details{
- System version (OS, version, user, working directory)

- R version

- Package names and versions

- Top-level script name and modification time

- Git hash, status and tag (if any; requires \code{\link{git2r}})
}
\examples{
 env_doc("print") # print information to stdout
 info <- env_doc() # return information as a consolidated data frame
}

