% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_environment.R
\name{env_doc}
\alias{env_doc}
\title{Environment info}
\usage{
env_doc(output = c("return", "print"), system = TRUE, version = TRUE,
  packages = TRUE, script = TRUE)
}
\arguments{
\item{output}{How should output be handled? return: return as a data frame (default);
print: print to stdout
@param system Include OS info from get_sysinfo()? Default TRUE
@param version Include R version?  Default TRUE
@param packages Include packages with repository and version from get_packageinfo()? Default TRUE
@param script Include script path and modification time from get_scriptinfo()? Default TRUE

@examples
env_doc("print") # print information to stdout
info <- env_doc() # return information as a consolidated data frame}
}
\description{
Document the working environment for a script including:
}
\details{
- System version (OS, version, user, working directory)
- R version
- Package names and versions
- Top-level script name and modification time
}

