\name{envDocument-package}
\alias{envDocument-package}
\alias{envDocument}
\docType{package}
\title{
document_environment: R package to print information of the R environment
}
\description{
Prints relevant information on the environment that an R script is running under
for use in reproducing results.  The report includes system, user, working directory
and included packages with versions.
}

\examples{
library(document_environment)
env_doc()
}

\details{
\tabular{ll}{
Package: \tab envDocument\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2015-04-02\cr
License: \tab Bristol-Myers Squibb Internal Use\cr
}
}
\author{
Donald Jackson
Bristol-Myers Squibb

donald.jackson@bms.com
}
\references{
}
\keyword{ package
}
\seealso{
}
