% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gitInfo.R
\name{get_gitInfo}
\alias{get_gitInfo}
\title{get_gitInfo}
\usage{
get_gitInfo(scriptPath = "")
}
\arguments{
\item{script}{Path to script (optional, defaults to get_scriptPath())}
}
\description{
Get git information from repository
}
\examples{
git_info <- get_gitinfo()
}

