% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/document_environment.R
=======
% Please edit documentation in R/get_packageinfo.R
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae
\name{get_packageinfo}
\alias{get_packageinfo}
\title{Get information on attached packages}
\usage{
get_packageinfo()
}
\description{
\code{get_packageinfo} returns information on attached packages
including name, version and source.
}
\examples{
packages <- get_packageinfo()
}

