% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packageinfo.R
\name{get_packageinfo}
\alias{get_packageinfo}
\title{Get information on attached packages}
\usage{
get_packageinfo()
}
\description{
\code{get_packageinfo} returns information on attached packages
including name, version and source.
}
\examples{
packages <- get_packageinfo()
}

