% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packageinfo.R
\name{get_packageinfo}
\alias{get_packageinfo}
\title{get_packageinfo}
\usage{
get_packageinfo()
}
\description{
Get information on attached packages, return as a data frame
}
\examples{
packages <- get_packageinfo()
}

