% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/document_environment.R
=======
% Please edit documentation in R/get_rversion.R
>>>>>>> 9279780d3e0173597aaee50f644bca2d297b91ae
\name{get_rversion}
\alias{get_rversion}
\title{Get  information on R version}
\usage{
get_rversion()
}
\description{
Get  information on R version
}
\examples{
Rversion <- get_rversion()
}

