% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_environment.R
\name{get_scriptpath}
\alias{get_scriptpath}
\title{get_scriptpath}
\usage{
get_scriptpath()
}
\description{
Get the path of the calling script
}
\details{
Returns the full path of the script that called this function (if any)
or NULL if path is not available
}
\examples{
mypath <- get_scriptpath()
}

