% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scriptpath.R
\name{get_scriptpath}
\alias{get_scriptpath}
\title{Get the path of the calling script}
\usage{
get_scriptpath(absolute = TRUE)
}
\arguments{
\item{absolute}{Return absolute path (absolute = TRUE, default) or path relative
to working directory}
}
\description{
\code{get_scriptpath} returns the full path of the script that called this function (if any)
or NULL if path is not available
}
\examples{
mypath <- get_scriptpath()

}

