% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/document_environment.R
\name{get_sysinfo}
\alias{get_sysinfo}
\title{Get system information}
\usage{
get_sysinfo()
}
\arguments{
\item{None}{}
}
\description{
Get system information: OS, hostname, userid, working directory as a data frame
}
\examples{
get_sysinfo()
}

