% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_environment.R
\name{get_sysinfo}
\alias{get_sysinfo}
\title{Get system information
  
Get system information: OS, hostname, userid, working directory as a data frame}
\usage{
get_sysinfo()
}
\description{
Get system information
  
Get system information: OS, hostname, userid, working directory as a data frame
}
\examples{
get_sysinfo()
}

