% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sysinfo.R
\name{get_sysinfo}
\alias{get_sysinfo}
\title{Get system information
  
Get OS, hostname, userid and working directory as a data frame}
\usage{
get_sysinfo()
}
\description{
Get system information
  
Get OS, hostname, userid and working directory as a data frame
}
\examples{
get_sysinfo()
}

