% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sysinfo.R
\name{get_sysinfo}
\alias{get_sysinfo}
\title{Get system information.}
\usage{
get_sysinfo()
}
\description{
\code{get_sysinfo} returns the OS, hostname, userid and working directory
 for the current analysis as a data frame
}
\examples{
get_sysinfo()

}

