% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominoInfo.R
\name{getDominoInfo}
\alias{getDominoInfo}
\title{Get information about a Domino Datalab run from environment variables in the run environment}
\usage{
getDominoInfo(drop_vars = c("DOMINO_API_HOST",
  "DOMINO_EXECUTOR_HOSTNAME", "DOMINO_USER_API_KEY"))
}
\arguments{
\item{drop_vars}{Variables that will not be included in results.  Character vector;
see \url{https://support.dominodatalab.com/hc/en-us/articles/205536355-Domino-environment-variables}
for a list of variables.}
}
\value{
data frame of Domino environment variables (renamed) and their values
}
\description{
Get information about a Domino Datalab run from environment variables in the run environment
}
