% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepo.R
\name{getRepo}
\alias{getRepo}
\title{getRepo}
\usage{
getRepo(testPath)
}
\arguments{
\item{testPath}{Path to a file to find/check repo}
}
\description{
Get git repository from path; makes sure path is tracked in repo.
}

