% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScriptInfo.R
\name{getScriptInfo}
\alias{getScriptInfo}
\title{Get information about the  calling script}
\usage{
getScriptInfo()
}
\description{
\code{getScriptInfo} retrieves the  file path and  modification time
for the calling script.
}
\details{
Note:
Currently this works if the script was called via source(),
knitr::spin(),  Rstudio's _compile notebook_ (ctrl-shift-K), 
Rscript or R -f.
}
\examples{
script_info <- getScriptInfo()
}
