% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyPrintInfo.R
\name{prettyPrintInfo}
\alias{prettyPrintInfo}
\title{Output environment information as individual tables per section}
\usage{
prettyPrintInfo(info_list = list())
}
\arguments{
\item{info_list}{R list with environment information}
}
\description{
Uses \code{knitr::kable} to pretty-print separate tables for each
'section' (list element) in the enviromnent information.
}
